fix(P7-final): CI-05 SHA pins (38 refs, 17 workflows) + CI-06 RUSTSEC stale ignores + SIM-02/SEL-04/WEB3-08/RELAY-05/SIM-07/11 docs + DeployTestnet mainnet guard + DeterministicFactory design note + manifest cleanup
Batch P7 final del workspace-extreme-audit-2026-09-24:
- CI-05: 38 action refs pineadas por SHA en 17 workflows
- CI-06: RUSTSEC-2024-0363 (sqlx 0.7.4→0.8.1) + RUSTSEC-2026-0189 (rmcp 0.3.2→3.1.4) ignores stale eliminados
- SIM-02: retained_spread GROSS documentado inline
- SEL-04: shim threshold divergencia documentada
- WEB3-08: factory permissionless documentado como design
- RELAY-05: parse_rpc_u128 limitación documentada
- SIM-07: intermediate_amount_out siempre None documentado
- SIM-11: SpecId::OSAKA fijo documentado
- WEB3-03: DeployTestnet rechaza chainid==1 (mainnet guard)
- PLAN_NORMALIDAD_E2E.md: auditoría de estado + plan de normalidad completo
- Manifests de auditoría (17k líneas) limpiados del staging

Verificación: cargo check --workspace ✓, 1356+135+75+123 tests ✓, tsc 3 proyectos ✓

+++ backend/searcher-rs/src/rhai_agent_bridge.rs
@@ -45,10 +45,10 @@ impl Usd {
         Self(BigDecimal::from(0))
     }
     pub fn is_positive(&self) -> bool {
-        self.0 > BigDecimal::from(0)
+        self.0 > 0
     }
     pub fn is_negative(&self) -> bool {
-        self.0 < BigDecimal::from(0)
+        self.0 < 0
     }
     pub fn checked_add(&self, rhs: &Self) -> Result<Self, String> {
         Ok(Self(&self.0 + &rhs.0))
